Correções flag '.'

+++ ft_printf/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: rdutenke <rdutenke@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/06 17:21:35 by rdutenke          #+#    #+#             */
-/*   Updated: 2020/05/25 13:51:30 by rdutenke         ###   ########.fr       */
+/*   Updated: 2020/05/26 14:54:02 by rdutenke         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@
 typedef struct	s_params
 {
 	va_list		ap;
-	size_t		len_c;
+	int		len_c;
 	char		*flag;
 	char		*str;
 	char		*v_str;
@@ -38,7 +38,7 @@ typedef struct	s_params
 	intmax_t	v_int;
 	char		conversion;
 	size_t			width;
-	size_t		precision;
+	int		precision;
 	int			len;
 	int			slen;
 	int 		i;
@@ -57,5 +57,6 @@ void	ft_print_char(t_ptf *parms);
 void	ft_print_str(t_ptf *parms);
 int		ft_match(char c, char *conversion);
 void	ft_padding(char type, t_ptf *parms);
-char    *ft_itoa_hex(uint32_t nb, char* base);
+char	*ft_itoa_hex(uint32_t nb, char* base);
+// void	ft_print_pecision(t_ptf *parms);
 #endif
